// ts/enums.d.ts
export const enum CardType {
	MINION = 4,
	INVALID = 0,
	HERO = 3,
	TOKEN = 9,
	SPELL = 5,
	ENCHANTMENT = 6,
	HERO_POWER = 10,
	ABILITY = 5,
	PLAYER = 2,
	ITEM = 8,
	GAME = 1,
	WEAPON = 7
}
// ...
export const enum CardClass {
	DEATHKNIGHT = 1,
	INVALID = 0,
	PRIEST = 6,
	NEUTRAL = 0,
	WARLOCK = 9,
	ROGUE = 7,
	COUNT = 12,
	DRUID = 2,
	SHAMAN = 8,
	MAGE = 4,
	WARRIOR = 10,
	PALADIN = 5,
	HUNTER = 3,
	DREAM = 11
}
// ...
export const enum Zone {
	DISCARD = -2,
	INVALID = 0,
	SETASIDE = 6,
	HAND = 3,
	DECK = 2,
	REMOVEDFROMGAME = 5,
	PLAY = 1,
	GRAVEYARD = 4,
	SECRET = 7
}